Moved `normal_pdf()` to `core_tools.py`
