+++ src/spikeinterface/core/__init__.py
@@ -88,6 +88,7 @@ from .globals import (
 from .core_tools import (
     read_python,
     write_python,
+    normal_pdf,
 )
 from .job_tools import ensure_n_jobs, ensure_chunk_size, ChunkRecordingExecutor, split_job_kwargs, fix_job_kwargs
 from .recording_tools import (
