+++ src/spikeinterface/core/core_tools.py
@@ -453,3 +453,26 @@ def is_editable_mode() -> bool:
     import spikeinterface
 
     return (Path(spikeinterface.__file__).parents[2] / "README.md").exists()
+
+
+def normal_pdf(x, mu: float = 0.0, sigma: float = 1.0):
+    """
+    Manual implementation of the Normal distribution pdf (probability density function).
+    It is about 8 to 10 times faster than scipy.stats.norm.pdf().
+
+    Parameters
+    ----------
+    x: scalar or array
+        The x-axis
+    mu: float, default: 0.0
+        The mean of the Normal distribution.
+    sigma: float, default: 1.0
+        The standard deviation of the Normal distribution.
+
+    Returns
+    -------
+    normal_pdf: scalar or array (same type as 'x')
+        The pdf of the Normal distribution for the given x-axis.
+    """
+
+    return 1 / (sigma * np.sqrt(2 * np.pi)) * np.exp(-((x - mu) ** 2) / (2 * sigma**2))
